add backup program
USE THIS if the version with the map doesn't function. This is the most
recent functional version of the project, leaving out the parts that
break the program. Demo the map creation wizard separately.

Currently only supports:

- start phase: selecting players, loading a map. both are for
demonstration purposes only: the actual playerlist and map that are used
in the program are hard-coded.

- phase ordering works: easy way to demonstrate is just hit enter and
watch everything go by

- fortification works. only works with country names c0-c4, ownership is
printed at the start of the program.